move generators to a dir
