+++ midi_tokenizers/quantized_midi_tokenizer.py
@@ -1,8 +1,8 @@
 import itertools
 
 import pandas as pd
-
 from quantizer_generator import QuantizerGenerator
+
 from midi_tokenizers.midi_tokenizer import MidiTokenizer
