+++ object_generators/tokenizer_generator.py
@@ -4,7 +4,7 @@ import streamlit as st
 
 from midi_tokenizers.midi_tokenizer import MidiTokenizer
 from midi_trainable_tokenizers.bpe_tokenizer import BpeTokenizer
-from base_tokenizer_generator import TokenizerFactory, BaseTokenizerGenerator
+from object_generators.base_tokenizer_generator import TokenizerFactory, BaseTokenizerGenerator
 
 
 class BpeTokenizerFactory(TokenizerFactory):
